Add sass and bootstrap support
