+++ frontend/package.json
@@ -11,9 +11,11 @@
   "author": "Mark Harris",
   "license": "ISC",
   "dependencies": {
+    "bootstrap": "^5.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.3.0"
+    "react-router-dom": "^6.3.0",
+    "sass": "^1.50.0"
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
@@ -25,6 +27,7 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-react": "^7.29.4",
     "html-webpack-plugin": "^5.5.0",
+    "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
     "webpack": "^5.69.1",
     "webpack-cli": "^4.9.2",
