+++ frontend/src/App.scss
@@ -1,0 +1,1 @@
+@import "../node_modules/bootstrap/scss/bootstrap";
